curaitor-agent/pyproject.toml
[tool]
name = "curaitor-agent"
version = "0.1.0"
description = "A project for downloading PDFs from arXiv, extracting text, and querying using an AI agent."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [
    "docling",
    "docling-core",
    "pandas",
    "pillow",
    "langchain",
    "python-dotenv",
    "requests",
    "PyPDF2",
    "beautifulsoup4",
]

[tool.poetry.scripts]
download_arxiv = "scripts.download_arxiv:main"
extract_and_index = "scripts.extract_and_index:main"
ask = "scripts.ask:main"